Rename to NavigationCounter
