+++ home/javascript/react/state/NavigationCounter.js
@@ -1,0 +1,16 @@
+
+let navigationCounts = {};
+let listeners = {};
+
+const subscribe = (keyName, callback) => {
+    const currentListeners = listeners[keyName] || [];
+    listeners[keyName] = [...currentListeners, callback];
+};
+
+const setCounts = (keyName, count) => {
+    navigationCounts[keyName] = count;
+    const currentListeners = listeners[keyName] || [];
+    currentListeners.forEach(listener => listener(count));
+};
+
+export { subscribe, setCounts };
